requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
enum34==1.1.6
idna==2.7
ipaddress==1.0.22
pycparser==2.19
PyJWT==1.6.4
requests==2.20.1
six==1.11.0
urllib3==1.24.1
